Updated mediawiki/skins Project: mediawiki/skins/BlueSky  ba7ff3d58cfeb3622b59a07ea38b70ec6689f54c
Integrating wikiHow's FollowWidget extension.

When [[MediaWiki:Bluesky-follow-table]] contains some HTML, this HTML is shown on the right side of the page (on LTR interfaces; the opposite for RTL).
For optimal results, the table should have id="follow_table" and follow certain other conventions, which will be documented in-depth on MediaWiki in the future.

Change-Id: I6fd4f756e13d028d172b7d47150ef47f5dcddc75