Updated mediawiki/skins Project: mediawiki/skins/BlueSky  abd6377c57c68221739ae15295e0d055390c633e
Integrated the UserLoginBox extension by Scott Cushman.

The box is shown as an item on the top menu on every page for anonymous
users, therefore allowing them to log in quickly without having to
navigate to the login page manually.

There are lots of things to be done here, but it's a start in any case.

Change-Id: Ife4de2526ef8e6d4ec3741d981ab6a9fce50e329