Updated mediawiki/skins Project: mediawiki/skins/erudite  e860ae68b78fa1d2acbef5baa1eb0fc48fa113f1
Fix position of list item bullets in footer

Since the footer div is set to align text to the right,
the text of the individual items is placed (naturally)
at the right edge of the page. However, in Google Chrome
the corresponding bullets aren't (they remain on the left).
To fix this one needs to add list-style-position: inside;
Tested with Chrome v37 and Firefox v32.

Change-Id: I3bb51f25ee247b9b10ad0d149b423ec80c59720e