Updated mediawiki/skins Project: mediawiki/skins/Nimbus  96ff0e64f1140d0d900b26c97a4f583b83673d5d
Removing CSS that duplicates the Comments extension's main CSS file

Roughly seven years ago certain skins and extensions were very tightly
coupled together, but luckily this isn't the case anymore. Comments has
been self-contained for a fair while and basically all of this CSS is in
/extensions/Comments/Comments.css, so the only rule we need is the one
that differs from the rules present in that file.

Change-Id: I5b2085d487cddf04371440f917e20674fcbc48e6