Updated mediawiki/skins Project: mediawiki/skins/Vector  b02d0da6a76843a9052189636a5086159e8a1b8f
Add version check for Vector

Sometimes, someone tries to install git master version of Vector on
a stable version of MediaWiki, which can result in a PHP error, e.g.
because of the missing status indicators in MW 1.24.

Add a version check to Vector.php to check, if a minimum mediawiki version
(actual 1.25) is installed in order to use this version.

Change-Id: I10f071253e48e6492949ebbc3c07964c747e885f