Updated mediawiki/skins Project: mediawiki/skins/Nostalgia  124da3017fedb126e00667b2c575d27acfee50c8
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I93612feffe1fd60989b18bcb43ae31d2e1017c15