Updated mediawiki/skins Project: mediawiki/skins/Vector  d4fef458afabed83ab0ba45d02f3c91b67bdac85
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: Ic3d188fe5a833a97027a06cd62c4acef691669bf