Updated mediawiki/skins Project: mediawiki/skins/MonoBook  19406c60b8d51ea6be2e1225625da91c2644e8cc
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I9d5cc55a5218bef618878e90773a99047d942beb