Updated mediawiki/skins Project: mediawiki/skins/CologneBlue  6d45639fe325fc3c8fbb5811dbbef21baca9f051
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I3103b2285c6ac3e9410ee9b62ca4c400db0a2227