Updated mediawiki/skins Project: mediawiki/skins/Modern  7fad18c6d1be062d413300de39a7e00662364554
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I05048fd3864ec918d365e5c7cbd68dcec78d4ea3