Updated mediawiki/skins Project: mediawiki/skins/chameleon  66e0d641567ebca47ace9ee8db252ff7ea141a86
Use FormEdit instead of normal Edit action if $sfgRenameEditTabs is set

SemanticForms just replaces the link title of the Edit action, not the action
ID in the page tools array structure. So, depending on $sfgRenameEditTabs
either the 'edit' action has to be used for the edit link or the 'form_edit'
action.

Change-Id: I0cb651bbd462821138c90cd5f43f93fc809b88c0