Updated mediawiki/skins Project: mediawiki/skins/Metrolook  c805474026f3627f76fcc79c22e68b7873184ef8
Fix last HTML validation issue

Swapped deprecated <br clear="all" /> to a div with the visualClear class.
MonoBook & friends use that class, too, so it should be a safe bet.

The output is *not* 1:1 identical with the previous code; the difference
is some pixels (~5-10px), but in fact the new version looks better to me
(based on a few quick tests), since now the gap between the category bar
and "last modified on X at Y" text is now smaller.

Change-Id: Ib75bd26400487fdd5d9b4f05bff9c3e0749e156a