Updated mediawiki/skins Project: mediawiki/skins/Metrolook  380f7f904cad8710c7f6952e0ae6976b59f6157f
Added two hooks for ad integration

An additional extension, such as [[mw:Extension:ShoutWiki Ads]], can
contain all the actual logic and CSS for displaying ads, but we still need
a few hook points here so that we can insert the ads in the correct
locations.

Technically speaking I guess even the #mw-panel-right HTML could be
offloaded to the advertising extension, but I don't see an empty div doing
much harm to anyone.

An example implementation is PidgiWiki
(http://www.pidgi.net/wiki/Main_Page), which is actually running an older
version of Metrolook. We don't want to bundle any ad code with Metrolook
itself, as such code would likely be site-specific as well as outside the
scope of what a skin should provide.

Change-Id: I7d93d304776b1f8c5bf274e89b29a9988d570fc2