Updated mediawiki/skins Project: mediawiki/skins/Gamepress  3af5154addf6047c875a0679804688a3196aa553
Allow the top menu to be used on mobile devices

Previously it wouldn't have been possible to "expand" a top menu item to
make its sub-menus visible on a mobile device, because tapping on a top
menu item would take you immediately to the target page instead of
expanding the menu.

Code by Bizzeebeever, I'm just committing it.

Change-Id: I2efb4b5ffd211f076b13b15fd6e4272e8fc1ee1f