Updated mediawiki/skins Project: mediawiki/skins/Example  07088ce4b56688e86533fc853a31c8428582bb62
Add "class": "ResourceLoaderSkinModule" with no explanation whatsoever.

Adds support for MonoBook/Vector style logos using wgLogo and wgLogoHD
(for HiDPI support), and possibly other things. It is also included in
mediawiki.skinning.interface, and thus probably redundant for most uses,
but some skins will need it and there is no documentation anywhere and we
might as well have it so stuff magically works.

Change-Id: I7149d16edc4b5d0a38269569ab1658c4d79c7eda