Updated mediawiki/skins Project: mediawiki/skins/MonoBook  762b25a551e315c0215b92c8fe8c0c3c80c26175
Show "Not logged in" message

Show "Not logged in" message left (or right depending on language) of an anon
user's personal tools in case general edit permission is given.

The display of the message was previously done in MW core
(SkinTemplate::buildPersonalUrls, introduced by I6f44e3e5d), which needs to be
reverted first (done in I5cfa9b4e).

Bug: T127758
Depends-on: I5cfa9b4e
Change-Id: I7001e311a8d5ed70129500ab1a3e99369d4a1d9d